back-end/requirements.txt
asgiref==3.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distlib==0.3.9
Django==5.1.3
django-taggit==6.1.0
djangorestframework==3.16.1
filelock==3.16.1
fonttools==4.59.2
idna==3.10
iniconfig==2.0.0
kiwisolver==1.4.9
Markdown==3.7
matplotlib==3.10.6
narwhals==2.14.0
node==1.2.2
numpy==2.2.3
numpy-financial==1.0.0
odict==1.9.0
package_name==0.1
packaging==24.2
pandas==2.3.2
pillow==11.1.0
platformdirs==4.3.6
plotly==6.5.0
pluggy==1.5.0
plumber==1.7
psycopg2-binary==2.9.10
pyparsing==3.2.3
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.2
sv_ttk==2.6.1
tzdata==2024.2
urllib3==2.5.0
virtualenv==20.28.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
